refactor: restructure of MetaMain component
