--- src/components/Asset/AssetContent/MetaMain.module.css
@@ -1,51 +1,0 @@
-.meta {
-  margin-bottom: calc(var(--spacer) / 1.5);
-  color: var(--color-secondary);
-  font-size: var(--font-size-small);
-}
-
-.asset {
-  margin-left: -2rem;
-  margin-right: -2rem;
-  padding-left: 2rem;
-  padding-right: 3rem;
-  border-bottom: 1px solid var(--border-color);
-  margin-bottom: calc(var(--spacer) / 1.5);
-  padding-bottom: calc(var(--spacer) / 1.75);
-}
-
-@media (min-width: 40rem) {
-  .asset {
-    margin-top: -0.65rem;
-  }
-}
-
-.assetType {
-  display: inline-block;
-  border-right: 1px solid var(--border-color);
-  padding-right: calc(var(--spacer) / 3.5);
-  margin-right: calc(var(--spacer) / 4);
-}
-
-.datatoken {
-  white-space: pre;
-  margin-right: calc(var(--spacer) / 3);
-}
-
-.byline {
-  font-size: var(--font-size-small);
-}
-
-.updated {
-  font-size: var(--font-size-mini);
-}
-
-.addWrap {
-  padding-left: calc(var(--spacer) / 5);
-  border-left: 1px solid var(--border-color);
-  display: inline-block;
-}
-
-.add {
-  font-size: var(--font-size-mini);
-}
